+++ NEWS.md
@@ -1,5 +1,10 @@
 # reqres (development version)
 
+* Add support for otel OpenTelemetry by creating a span along with a request and
+  close it once done. spans created during the handling of the request should
+  use this span as a parent. Further, support the metrics defined in the HTTP
+  semantics
+
 # reqres 1.0.0
 
 * Use rlang native type checking instead of assertthat